+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 - Important fix to support GLOSSES with a dash (-) in their name.
 - main.py can now be invoked also from another directory: `player` module and `assets` are searched locally.
+- Important option changed `--corpus-generated-directory` --> `--dictionary-dir`
 
 ## [0.2.0] - 2025-08-18